+++ apps/consumer-client/pages/_app.tsx
@@ -1,0 +1,15 @@
+import { GlobalStyle } from "../components/GlobalStyle";
+
+/**
+ * This component wraps all pages in this Next.js application.
+ */
+const App = ({ Component, pageProps }: { Component: any; pageProps: any }) => {
+  return (
+    <>
+      <GlobalStyle />
+      <Component {...pageProps} />
+    </>
+  );
+};
+
+export default App;