+++ apps/consumer-client/src/util.ts
@@ -1,13 +1,3 @@
-export const IS_PROD = process.env.NODE_ENV === "production";
-
-export const PASSPORT_URL = IS_PROD
-  ? "https://zupass.org/"
-  : "http://localhost:3000/";
-
-export const PASSPORT_SERVER_URL = IS_PROD
-  ? "https://api.pcd-passport.com/"
-  : "http://localhost:3002/";
-
 // Popup window will redirect to the passport to request a proof.
 // Open the popup window under the current domain, let it redirect there:
 export function requestProofFromPassport(proofUrl: string) {