+++ packages/passport-interface/src/PCDIntegration.ts
@@ -21,3 +21,25 @@ export function useProof<T extends PCDPackage>(
 
   return proof;
 }
+
+/**
+ * React hook that listens for PCDs returned by the passport to the application.
+ */
+export function usePassportPCD() {
+  const [pcdStr, setPcdStr] = useState("");
+
+  // Listen for PCDs coming back from the Passport popup
+  useEffect(() => {
+    function receiveMessage(ev: MessageEvent<any>) {
+      // This next line is important. Extensions including Metamask apparently
+      // send messages to every page. Ignore those.
+      if (!ev.data.encodedPcd) return;
+      console.log("Received message", ev.data);
+      setPcdStr(ev.data.encodedPcd);
+    }
+    window.addEventListener("message", receiveMessage, false);
+    return () => window.removeEventListener("message", receiveMessage);
+  }, []);
+
+  return pcdStr;
+}