+++ packages/passport-interface/src/PassportInterface.ts
@@ -10,11 +10,19 @@ export interface PCDRequest {
   type: PCDRequestType;
 }
 
+export interface GetRequestOptions {
+  genericProveScreen?: boolean;
+  title?: string;
+  description?: string;
+  debug?: boolean;
+}
+
 export interface PCDGetRequest<T extends PCDPackage = PCDPackage>
   extends PCDRequest {
   type: PCDRequestType.Get;
   pcdType: T["name"];
   args: ArgsOf<T>;
+  options?: GetRequestOptions;
 }
 
 export interface PCDAddRequest extends PCDRequest {
@@ -26,13 +34,15 @@ export function constructPassportPcdGetRequestUrl<T extends PCDPackage>(
   passportOrigin: string,
   returnUrl: string,
   pcdType: T["name"],
-  args: ArgsOf<T>
+  args: ArgsOf<T>,
+  options?: GetRequestOptions
 ) {
   const req: PCDGetRequest<T> = {
     type: PCDRequestType.Get,
     returnUrl: returnUrl,
     args: args,
     pcdType,
+    options,
   };
   const encReq = encodeURIComponent(JSON.stringify(req));
   return `${passportOrigin}#/prove?request=${encReq}`;