+++ packages/eslint-config-custom/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @pcd/eslint-config-custom
 
+## 0.0.8
+
+### Patch Changes
+
+- Implemented ProofOptions interface.
+
 ## 0.0.7
 
 ### Patch Changes